+++ build-logic/convention/build.gradle.kts
@@ -14,14 +14,15 @@ dependencies {
 
 gradlePlugin {
     plugins {
-        libs.plugins.cleanarchitecturemapper.apply {
+        libs.plugins.cleanwizard.apply {
 
             val pluginConfigVersions = pluginConfig.versions
 
-            register(libs.plugins.cleanwizard.pluginId) {
-                id = libs.plugins.cleanwizard.pluginId
+            register(core.pluginId) {
+                id = core.pluginId
                 implementationClass = pluginConfigVersions.cleanwizard.implementation.get()
             }
+
             register(kotlin.pluginId) {
                 id = kotlin.pluginId
                 implementationClass = pluginConfigVersions.foundation.kotlin.implementation.get()
