+++ build-logic/convention/src/main/kotlin/corp/tbm/cleanwizard/buildLogic/convention/foundation/CleanWizardJsonSerializer.kt
@@ -1,4 +1,4 @@
-package corp.tbm.cleanarchitecturemapper.buildLogic.convention.foundation
+package corp.tbm.cleanwizard.buildLogic.convention.foundation
 
 enum class CleanWizardJsonSerializer(val serializer: String) {
     KOTLINX_SERIALIZATION("kotlinx-serialization"),
