+++ build-logic/convention/src/main/kotlin/corp/tbm/cleanwizard/buildLogic/convention/foundation/CleanWizardProcessorConfig.kt
@@ -1,4 +1,4 @@
-package corp.tbm.cleanarchitecturemapper.buildLogic.convention.foundation
+package corp.tbm.cleanwizard.buildLogic.convention.foundation
 
 open class CleanWizardProcessorConfig(
     var dataModuleName: String = "data",
