+++ build-logic/convention/src/main/kotlin/corp/tbm/cleanwizard/buildLogic/convention/plugins/codegen/CodegenPlugin.kt
@@ -1,8 +1,8 @@
-package corp.tbm.cleanarchitecturemapper.buildLogic.convention.plugins.codegen
+package corp.tbm.cleanwizard.buildLogic.convention.plugins.codegen
 
-import corp.tbm.cleanarchitecturemapper.buildLogic.convention.foundation.extensions.applyPlugin
-import corp.tbm.cleanarchitecturemapper.buildLogic.convention.foundation.extensions.implementation
-import corp.tbm.cleanarchitecturemapper.buildLogic.convention.foundation.extensions.libs
+import corp.tbm.cleanwizard.buildLogic.convention.foundation.extensions.applyPlugin
+import corp.tbm.cleanwizard.buildLogic.convention.foundation.extensions.implementation
+import corp.tbm.cleanwizard.buildLogic.convention.foundation.extensions.libs
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.dependencies
@@ -13,7 +13,7 @@ internal class CodegenPlugin : Plugin<Project> {
     override fun apply(target: Project) {
         with(target) {
 
-            applyPlugin(libs.plugins.cleanarchitecturemapper.kotlin)
+            applyPlugin(libs.plugins.cleanwizard.kotlin)
             applyPlugin(libs.plugins.google.devtools.ksp)
 
             dependencies {
