+++ build-logic/convention/src/main/kotlin/corp/tbm/cleanwizard/buildLogic/convention/plugins/codegen/VisitorPlugin.kt
@@ -1,8 +1,8 @@
-package corp.tbm.cleanarchitecturemapper.buildLogic.convention.plugins.codegen
+package corp.tbm.cleanwizard.buildLogic.convention.plugins.codegen
 
-import corp.tbm.cleanarchitecturemapper.buildLogic.convention.foundation.extensions.applyPlugin
-import corp.tbm.cleanarchitecturemapper.buildLogic.convention.foundation.extensions.implementation
-import corp.tbm.cleanarchitecturemapper.buildLogic.convention.foundation.extensions.libs
+import corp.tbm.cleanwizard.buildLogic.convention.foundation.extensions.applyPlugin
+import corp.tbm.cleanwizard.buildLogic.convention.foundation.extensions.implementation
+import corp.tbm.cleanwizard.buildLogic.convention.foundation.extensions.libs
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.dependencies
@@ -12,7 +12,7 @@ internal class VisitorPlugin : Plugin<Project> {
     override fun apply(target: Project) {
         with(target) {
 
-            applyPlugin(libs.plugins.cleanarchitecturemapper.codegen.foundation)
+            applyPlugin(libs.plugins.cleanwizard.codegen.foundation)
 
             dependencies {
                 implementation(project(":foundation:codegen:universal"))
