+++ foundation/annotations/build.gradle.kts
@@ -1,3 +1,3 @@
 plugins {
-    alias(libs.plugins.cleanarchitecturemapper.kotlin)
+    alias(libs.plugins.cleanwizard.kotlin)
 }
