+++ foundation/codegen/kotlinpoet/build.gradle.kts
@@ -1,3 +1,3 @@
 plugins {
-    alias(libs.plugins.cleanarchitecturemapper.codegen.foundation)
+    alias(libs.plugins.cleanwizard.codegen.foundation)
 }
