+++ foundation/codegen/ksp/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    alias(libs.plugins.cleanarchitecturemapper.codegen.foundation)
+    alias(libs.plugins.cleanwizard.codegen.foundation)
 }
 dependencies {
     implementation(projects.foundation.codegen.universal)
