+++ foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/ModelType.kt
@@ -1,8 +1,8 @@
-package corp.tbm.cleanarchitecturemapper.foundation.codegen.universal
+package corp.tbm.cleanwizard.foundation.codegen.universal
 
-import corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.processor.ProcessorOptions.domainOptions
-import corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.processor.ProcessorOptions.dtoOptions
-import corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.processor.ProcessorOptions.uiOptions
+import corp.tbm.cleanwizard.foundation.codegen.universal.processor.ProcessorOptions.domainOptions
+import corp.tbm.cleanwizard.foundation.codegen.universal.processor.ProcessorOptions.dtoOptions
+import corp.tbm.cleanwizard.foundation.codegen.universal.processor.ProcessorOptions.uiOptions
 
 enum class ModelType(val suffix: String, val packageName: String) {
     DTO(dtoOptions.suffix, dtoOptions.packageName),
