+++ foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/exceptions/CleanWizardPropertyAlreadyMarkedWithEnumException.kt
@@ -1,3 +1,3 @@
-package corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.exceptions
+package corp.tbm.cleanwizard.foundation.codegen.universal.exceptions
 
 class CleanWizardFieldAlreadyMarkedWithEnumException(message: String) : Exception(message)
