+++ foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/exceptions/references/PropertyAlreadyMarkedWithEnum.kt
@@ -1,5 +1,5 @@
-package corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.exceptions.references
+package corp.tbm.cleanwizard.foundation.codegen.universal.exceptions.references
 
-import corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.exceptions.CleanWizardFieldAlreadyMarkedWithEnumException
+import corp.tbm.cleanwizard.foundation.codegen.universal.exceptions.CleanWizardFieldAlreadyMarkedWithEnumException
 
 typealias PropertyAlreadyMarkedWithEnumException = CleanWizardFieldAlreadyMarkedWithEnumException
