+++ foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/extensions/ksp/ResolverExtensions.kt
@@ -1,4 +1,4 @@
-package corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.extensions.ksp
+package corp.tbm.cleanwizard.foundation.codegen.universal.extensions.ksp
 
 import com.google.devtools.ksp.processing.Resolver
