+++ foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/extensions/ksp/ks/KSAnnotationExtensions.kt
@@ -1,7 +1,7 @@
-package corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.extensions.ksp.ks
+package corp.tbm.cleanwizard.foundation.codegen.universal.extensions.ksp.ks
 
 import com.google.devtools.ksp.symbol.KSAnnotation
-import corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.EnumType
+import corp.tbm.cleanwizard.foundation.codegen.universal.EnumType
 
 inline val KSAnnotation.name
     get() = shortName.asString()
