+++ foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/extensions/ksp/ks/KSTypeExtensions.kt
@@ -1,7 +1,7 @@
-package corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.extensions.ksp.ks
+package corp.tbm.cleanwizard.foundation.codegen.universal.extensions.ksp.ks
 
 import com.google.devtools.ksp.symbol.KSType
-import corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.processor.ProcessorOptions
+import corp.tbm.cleanwizard.foundation.codegen.universal.processor.ProcessorOptions
 
 inline val KSType.isListSubclass
     get() = declaration.qualifiedName?.asString()?.endsWith("List") == true
