+++ foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/extensions/ksp/ks/KSValueArgumentExtensions.kt
@@ -1,4 +1,4 @@
-package corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.extensions.ksp.ks
+package corp.tbm.cleanwizard.foundation.codegen.universal.extensions.ksp.ks
 
 import com.google.devtools.ksp.symbol.KSValueArgument
