+++ foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/processor/JsonSerializer.kt
@@ -1,4 +1,4 @@
-package corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.processor
+package corp.tbm.cleanwizard.foundation.codegen.universal.processor
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.google.gson.annotations.SerializedName
