+++ foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/processor/ProcessorOptions.kt
@@ -1,4 +1,4 @@
-package corp.tbm.cleanarchitecturemapper.foundation.codegen.universal.processor
+package corp.tbm.cleanwizard.foundation.codegen.universal.processor
 
 object ProcessorOptions {
