+++ processor/build.gradle.kts
@@ -1,7 +1,7 @@
 import com.vanniktech.maven.publish.SonatypeHost
 
 plugins {
-    alias(libs.plugins.cleanarchitecturemapper.codegen.visitor)
+    alias(libs.plugins.cleanwizard.codegen.visitor)
     alias(libs.plugins.vanniktech.maven.publish)
 }
