+++ processor/src/main/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -1,1 +1,1 @@
-corp.tbm.cleanarchitecturemapper.processor.DTOProcessorProvider
+corp.tbm.cleanwizard.processor.DTOProcessorProvider
